apps/api/requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pydantic==2.12.5
python-dotenv==1.1.1
PyJWT==2.10.1
